+++ console/src/components/Main/State.ts
@@ -17,6 +17,7 @@ export interface MainState {
   serverConfig: {
     data: {
       version: string;
+      is_function_permissions_inferred: boolean;
       is_admin_secret_set: boolean;
       is_auth_hook_set: boolean;
       is_jwt_set: boolean;
@@ -50,6 +51,7 @@ const defaultState: MainState = {
   serverConfig: {
     data: {
       version: '',
+      is_function_permissions_inferred: true,
       is_admin_secret_set: false,
       is_auth_hook_set: false,
       is_jwt_set: false,